Language.Haskell.Exts.Util

Free variables of ASTs

class FreeVars a where

Minimal complete definition

freeVars

Methods

freeVars :: a -> Set (Name ())

Return the variables, erring on the side of more free variables

data Vars

Constructors

Vars

Fields

bound :: Set (Name ())

free :: Set (Name ())

class AllVars a where

Minimal complete definition

allVars

Methods

allVars :: a -> Vars

Return the variables, erring on the side of more free variables

Rebracketing of ASTs

class Brackets a where

Minimal complete definition

remParen, addParen, isAtom, needBracket

Methods

remParen :: a -> Maybe a

addParen :: a -> a

isAtom :: a -> Bool

Is this item lexically requiring no bracketing ever i.e. is totally atomic

needBracket :: Int -> a -> a -> Bool

Is the child safe free from brackets in the parent position. Err on the side of caution, True = don't know

paren :: (Data l, Default l) => Exp l -> Exp l

Add a Paren around something if it is not atomic

transformBracket :: (Data l, Default l) => (Exp l -> Maybe (Exp l)) -> Exp l -> Exp l

rebracket1 :: (Data l, Default l) => Exp l -> Exp l

Add/remove brackets as suggested needBracket at 1-level of depth

appsBracket :: (Data l, Default l) => [Exp l] -> Exp l

Orphan instances

Default SrcLoc

Instance details

Methods

def :: SrcLoc

Default SrcSpan

Instance details

Methods

def :: SrcSpan

Default SrcSpanInfo

Instance details

Methods

def :: SrcSpanInfo